Updates application callback URIs into the gitlab charm relation
When performing juju refresh, the container will get recreated with
a newer juju charm version, which will mean that the container may
have a different IP than it had. That means that the callback URIs
written into the gitlab relation are stale, and we need to refresh that.

The same may happen any time the Kubernetes pod crashes.